+++ entrypoints/popup/App.vue
@@ -7,11 +7,13 @@ import { useI18n } from '@/utils/i18n'
 
 const { t } = useI18n()
 const isValidUrl = ref(false)
+const isLocalPDFFileUrl = ref(false)
 browser.tabs.query({ active: true, currentWindow: true }).then(async (tabs) => {
   if (tabs.length === 1) {
     const tab = tabs[0]
     const { url } = tab || {}
     isValidUrl.value = !!url && /https?:\/\//.test(url) && !INVALID_URLS.some((regexp) => regexp.test(url))
+    isLocalPDFFileUrl.value = !!url && /^file:\/\/.*\.pdf$/.test(url)
   }
 })
 </script>
@@ -20,6 +22,9 @@ browser.tabs.query({ active: true, currentWindow: true }).then(async (tabs) => {
   <div>
     <div v-if="isValidUrl">
       {{ t('popup.reload_page') }}
+    </div>
+    <div v-else-if="isLocalPDFFileUrl">
+      {{ t('popup.local_pdf_file_not_supported') }}
     </div>
     <div v-else>
       {{ t('popup.page_not_supported') }}