+++ utils/rpc/content-fns.ts
@@ -22,7 +22,7 @@ export type EventKey = keyof Events
 const parsePdfFileOfCurrentUrl = memoFunction(async (url: string) => {
   const resp = await fetch(url)
   const contentType = resp.headers.get('content-type')
-  if (contentType === 'application/pdf') {
+  if (contentType?.includes('application/pdf')) {
     let fileName = resp.headers.get('content-disposition')?.split('filename=')[1] ?? location.href.split('/').pop() ?? ''
     fileName = fileName.replace(/"/g, '')
     if (!fileName.endsWith('.pdf')) fileName += '.pdf'